+++ src/Facades/DisplayFilter.php
@@ -6,10 +6,10 @@ use Illuminate\Support\Facades\Facade;
 
 /**
  * Class Filter.
- * @method static \SleepingOwl\Admin\Display\Filter\FilterCustom custom($name)
- * @method static \SleepingOwl\Admin\Display\Filter\FilterField field($name)
- * @method static \SleepingOwl\Admin\Display\Filter\FilterRelated related($name)
- * @method static \SleepingOwl\Admin\Display\Filter\FilterScope scope($name)
+ * @method static \SleepingOwl\Admin\Display\Filter\FilterCustom custom($name, string|\Closure|null $title, \Closure $callback)
+ * @method static \SleepingOwl\Admin\Display\Filter\FilterField field($name, string|\Closure|null $title)
+ * @method static \SleepingOwl\Admin\Display\Filter\FilterRelated related($namem, string|\Closure|null $title)
+ * @method static \SleepingOwl\Admin\Display\Filter\FilterScope scope($name, string|\Closure|null $title)
  */
 class DisplayFilter extends Facade
 {
